+++ example/lib/main.dart
@@ -64,10 +64,110 @@ const Map<String, dynamic> STRING_REGEXP_CONST = {
   USER_ID_TAG_REGEXP_TYPE: USER_TAG_WITH_ID,
   ARABIC_REGEXP_TYPE: ARABIC_REGEXP,
 };
+List<DetectorOptions> detectorOptionsList = [
+  DetectorOptions(
+      type: "PHONE",
+      pattern: PHONE_REGEXP,
+      style: const TextStyle(
+        color: Colors.red,
+        fontSize: 24,
+      ),
+      onTap: (url) {
+        //launch("tel:" + url);
+        debugPrint("tel:${url.value}");
+      }),
+  DetectorOptions(
+    pattern: URL_REGEXP,
+    type: URL_REGEXP_TYPE,
+    style: const TextStyle(
+      fontSize: 16,
+      fontWeight: FontWeight.bold,
+      color: Colors.brown,
+    ),
+    onTap: (value) {
+      debugPrint("${value.detectType}:${value.value}:${value.regExp.pattern}");
+    },
+  ),
+  DetectorOptions(
+    pattern: EMAIL_REGEXP,
+    type: EMAIL_REGEXP_TYPE,
+    style: const TextStyle(
+      fontSize: 16,
+      fontWeight: FontWeight.bold,
+      color: Colors.purple,
+    ),
+    onTap: (value) {
+      debugPrint("${value.detectType}:${value.value}:${value.regExp.pattern}");
+    },
+  ),
+  DetectorOptions(
+    pattern: HASHTAG_REGEXP,
+    type: HASHTAG_REGEXP_TYPE,
+    style: const TextStyle(
+      fontSize: 16,
+      fontWeight: FontWeight.bold,
+      color: Colors.red,
+    ),
+    onTap: (value) {
+      debugPrint("${value.detectType}:${value.value}:${value.regExp.pattern}");
+    },
+  ),
+  DetectorOptions(
+    pattern: USER_TAG_REGEXP,
+    type: USER_TAG_REGEXP_TYPE,
+    style: const TextStyle(
+      fontSize: 16,
+      fontWeight: FontWeight.bold,
+      color: Colors.green,
+    ),
+    onTap: (value) {
+      debugPrint("${value.detectType}:${value.value}:${value.regExp.pattern}");
+    },
+  ),
+  DetectorOptions(
+    pattern: USER_TAG_WITH_ID,
+    type: USER_ID_TAG_REGEXP_TYPE,
+    style: const TextStyle(
+      fontSize: 16,
+      fontWeight: FontWeight.bold,
+      color: Colors.purpleAccent,
+    ),
+    onTap: (value) {
+      debugPrint("${value.detectType}:${value.value}:${value.regExp.pattern}");
+    },
+    parsingValue: (value) {
+      RegExp customRegExp = RegExp(r'''[a-zA-Z\u0600-\u06FF]+''');
+      Match? match = customRegExp.firstMatch(value.value);
+      DetectedValue val = value;
+      if (match != null) {
+        var ma = match[0].toString();
+        val = DetectedValue.init(ma, "NoD", customRegExp);
+      }
+      debugPrint("${value.detectType}:${value.value}:${value.regExp.pattern}");
+      debugPrint("${val.detectType}:${val.value}:${val.regExp.pattern}");
+      return val;
+    },
+    valueWidget: (value) {
+      return Container(
+        color: Colors.blue,
+        child: Text(
+          value.value,
+          style: const TextStyle(color: Colors.white),
+        ),
+      );
+    },
+  ),
+];
 
-class JTextDetectorExample extends StatelessWidget {
+class JTextDetectorExample extends StatefulWidget {
   const JTextDetectorExample({Key? key}) : super(key: key);
 
+  @override
+  State<JTextDetectorExample> createState() => _JTextDetectorExampleState();
+}
+
+class _JTextDetectorExampleState extends State<JTextDetectorExample> {
+  TextEditingController controller = TextEditingController(text: '');
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -79,6 +179,20 @@ class JTextDetectorExample extends StatelessWidget {
           Expanded(
               child: ListView(
             children: [
+              JTextFieldDetector(
+                controller: controller,
+                keyboardType: TextInputType.multiline,
+                style: const TextStyle(
+                  fontSize: 16,
+                  fontWeight: FontWeight.bold,
+                  color: Colors.black,
+                ),
+                textAlign: TextAlign.center,
+                textDirection: TextDirection.rtl,
+                clipBehavior: Clip.hardEdge,
+                detectorOptions: detectorOptionsList,
+                maxLines: 15,
+              ),
               for (int i = 0; i < TEXT_EXAMPLE.length; i++)
                 JTextDetector(
                   text: TEXT_EXAMPLE[i],