+++ lib/jtdetector.dart
@@ -1,7 +1,14 @@
 library jtdetector;
 
+import 'dart:ui' as ui show BoxHeightStyle, BoxWidthStyle;
+
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter/material.dart';
 
 part 'src/jHelpers.dart';
 part 'src/jTextDetector.dart';
+part 'src/jTextFieldDetector.dart';