+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.1.1
+* Text Field Detector Widget
+* Add Text Field Detector Example
 ## 1.1.0
 * Hot Fix
 ## 1.0.0